+++ tests/test_read.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from frmb._read import FrmbFormat
+from frmb._read import FrmbMenuItem
 from frmb._read import read_menu_hierarchy
 from frmb._read import read_menu_hierarchy_as_file
 from frmb._read import validate_menu_hierarchy
@@ -106,10 +106,10 @@ def test__validate_entry_hierarchy__show(data_dir):
 
 def test__validate_entry_hierarchy_recursion(data_dir):
 
-    hierarchy = FrmbFormat("lowest", "", None, tuple(), tuple(), tuple())
+    hierarchy = FrmbMenuItem("lowest", "", None, tuple(), tuple(), tuple())
     for i in range(16):
         p = tuple() if i < 15 else ("p",)
-        hierarchy = FrmbFormat(f"{i}", f"{i}", None, tuple(), p, (hierarchy,))
+        hierarchy = FrmbMenuItem(f"{i}", f"{i}", None, tuple(), p, (hierarchy,))
 
     errors, warnings = validate_menu_hierarchy([hierarchy])
     assert len(errors) == 1
@@ -117,27 +117,27 @@ def test__validate_entry_hierarchy_recursion(data_dir):
     assert len(warnings) == 0
 
 
-def test_FrmbFormat_hash():
+def test__FrmbMenuItem__hash():
 
-    hierarchy_1 = FrmbFormat("lowest", "", None, tuple(), tuple(), tuple())
+    hierarchy_1 = FrmbMenuItem("lowest", "", None, tuple(), tuple(), tuple())
     for i in range(13):
         path = tuple() if i < 12 else ("p",)
-        hierarchy_1 = FrmbFormat(f"{i}", f"{i}", None, tuple(), path, (hierarchy_1,))
+        hierarchy_1 = FrmbMenuItem(f"{i}", f"{i}", None, tuple(), path, (hierarchy_1,))
 
-    hierarchy_2 = FrmbFormat("lowest", "", None, tuple(), tuple(), tuple())
+    hierarchy_2 = FrmbMenuItem("lowest", "", None, tuple(), tuple(), tuple())
     for i in range(12):
         path = tuple() if i < 11 else ("p",)
-        hierarchy_2 = FrmbFormat(f"{i}", f"{i}", None, tuple(), path, (hierarchy_2,))
+        hierarchy_2 = FrmbMenuItem(f"{i}", f"{i}", None, tuple(), path, (hierarchy_2,))
 
-    hierarchy_3 = FrmbFormat("lowest", "", None, tuple(), tuple(), tuple())
+    hierarchy_3 = FrmbMenuItem("lowest", "", None, tuple(), tuple(), tuple())
     for i in range(12):
         path = tuple() if i < 11 else ("p",)
-        hierarchy_3 = FrmbFormat(f"{i}", f"{i}", None, tuple(), path, (hierarchy_3,))
+        hierarchy_3 = FrmbMenuItem(f"{i}", f"{i}", None, tuple(), path, (hierarchy_3,))
 
-    hierarchy_4 = FrmbFormat("lowest", "", None, tuple(), tuple(), tuple())
+    hierarchy_4 = FrmbMenuItem("lowest", "", None, tuple(), tuple(), tuple())
     for i in range(12):
         path = tuple() if i < 11 else ("p",)
-        hierarchy_4 = FrmbFormat(
+        hierarchy_4 = FrmbMenuItem(
             f"{i if i < 11 else 'babz'}", f"{i}", None, tuple(), path, (hierarchy_4,)
         )
 
@@ -146,7 +146,7 @@ def test_FrmbFormat_hash():
     assert hash(hierarchy_3) != hash(hierarchy_4)
 
 
-def test_FrmbFormat_tofile(data_dir, tmp_path):
+def test__FrmbMenuItem__tofile(data_dir, tmp_path):
     # ensure to_file use the same logic as from_file
     src_dir = data_dir / "structure1" / "studio"
     hierarchy_src = read_menu_hierarchy(src_dir)